replace /etc/default/named with a systemd drop-in file
systemd drop-in files work across distributions, unlike
/etc/default/named-style shell files. Also you can see the full config
with "systemctl cat" when using drop-in files.

Don't know why the content tests fail with this message: "The diff is
empty, are your objects producing identical `#inspect` output?"

Apparently that means that there are no diffs, but the Ruby objects
are not equal. Don't know what I can do about that though.